Split test code from lib code for dodgy_2d.
